backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pyswisseph==2.10.3.2
python-multipart==0.0.6
google-generativeai==0.8.6
python-dotenv==1.0.0
cryptography==41.0.7
python-dateutil==2.8.2
bcrypt==4.1.2
PyJWT==2.8.0
psutil==5.9.6
requests==2.31.0
twilio==8.10.0
pytz==2023.3
timezonefinder==6.2.0
google-cloud-storage==2.10.0
Pillow==10.0.1
replicate==0.15.4